FIX: Fixed recursion bug with status. Moved __repr__ implementatio to SndDevice. Removed PythonSignal in favor of AttributeSignal. Changed none_if_no_parent to be nan_if_no_parent
