+++ hxrsnd/tests/test_macromotor.py
@@ -15,6 +15,7 @@ logger = logging.getLogger(__name__)
 @pytest.mark.parametrize("dev", get_classes_in_module(macromotor, Device))
 def test_devices_instantiate_and_run_ophyd_functions(dev):
     device = fake_device(dev)
+    device
     assert(isinstance(device.read(), dict))
     assert(isinstance(device.describe(), dict))
     assert(isinstance(device.describe_configuration(), dict))
