+++ hxrsnd/tower.py
@@ -292,17 +292,6 @@ class TowerBase(SndDevice):
         else:
             return status
 
-    def __repr__(self):
-        """
-        Returns the status of the device. Alias for status().
-
-        Returns
-        -------
-        status : str
-            Status string.
-        """
-        return self.status(print_status=False)
-
         
 class DelayTower(TowerBase):
     """
